REF: update substrate management in Circuit and test cases
